@@ -3,13 +3,30 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
+    "start": "node server/index.js",
     "serve": "vue-cli-service serve",
     "build": "vue-cli-service build",
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
+    "@fortawesome/fontawesome-svg-core": "^6.1.1",
+    "@fortawesome/free-solid-svg-icons": "^6.1.1",
+    "@fortawesome/vue-fontawesome": "^3.0.0-5",
+    "@vuelidate/core": "^2.0.0-alpha.41",
+    "@vuelidate/validators": "^2.0.0-alpha.29",
+    "bulma": "^0.9.4",
     "core-js": "^3.8.3",
-    "vue": "^2.6.14"
+    "css-loader": "^5.2.7",
+    "express": "^4.18.1",
+    "firebase": "^9.8.1",
+    "sass": "^1.51.0",
+    "sass-loader": "^10.2.1",
+    "serve-static": "^1.15.0",
+    "slugify": "^1.6.5",
+    "vue": "^3.2.34",
+    "vue-router": "^4.0.15",
+    "vue-toastification": "^2.0.0-rc.5",
+    "vuex": "^4.0.2"
   },
   "devDependencies": {
     "@babel/core": "^7.12.16",
@@ -33,7 +50,10 @@
     "parserOptions": {
       "parser": "@babel/eslint-parser"
     },
-    "rules": {}
+    "rules": {
+      "no-debugger": 0,
+      "no-console": 0
+    }
   },
   "browserslist": [
     "> 1%",
