@@ -1,8 +1,27 @@
-import Vue from 'vue'
+import { createApp } from 'vue'
 import App from './App.vue'
+import router from './router'
+import store from './store'
+import Toast from 'vue-toastification';
+import 'vue-toastification/dist/index.css'
+import { library } from '@fortawesome/fontawesome-svg-core'
+import { faStar, faUpload } from '@fortawesome/free-solid-svg-icons'
+import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome'
 
-Vue.config.productionTip = false
+library.add([faStar, faUpload])
 
-new Vue({
-  render: h => h(App),
-}).$mount('#app')
+let app;
+store.dispatch('user/onAuthChange',() => {
+   if(!app){
+      app = createApp(App);
+
+      app.use(store)
+      .use(router)
+      .use(Toast);
+   
+      app.component("font-awesome-icon", FontAwesomeIcon);
+      app.mount('#app')
+   }
+
+});
+
